+++ packages/multichain/src/scope/constants.test.ts
@@ -53,6 +53,7 @@ describe('KnownRpcMethods', () => {
           "eth_syncing",
           "eth_uninstallFilter",
         ],
+        "solana": Array [],
       }
     `);
   });